Add ChatManager class to handle chat functionality
